requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
python-telegram-bot==20.7
python-dotenv==1.0.0
